Adding invoices to database -> Invoice is now added to database -> InvoiceMetaData is now added to database -> Fixed float numbers where in some case was shown with more than 2 decimal places -> Improved robustness when adding parts to invoice print preview

+++ src/main/java/com/mechanic/code/database/Part.java
@@ -7,7 +7,7 @@ import java.text.DecimalFormat;
 public class Part {
 	private int counter,invoiceID,quantity;
 	private String partsID,description;
-	private float price;
+	private double price;
 	private static DecimalFormat decimalFormat=new DecimalFormat("0.00");
 
 	public Part(){
@@ -19,7 +19,7 @@ public class Part {
 		description="";
 	}
 
-	public Part(int count,int arithmosTimologiou,int posotita,String onoma,String kodikos,float timi){
+	public Part(int count,int arithmosTimologiou,int posotita,String onoma,String kodikos,double timi){
 		invoiceID=arithmosTimologiou;
 		quantity=posotita;
 		partsID=kodikos;
